pyproject.toml
[tool.poetry]
name = "read-write-google-sheets-api"
version = "0.1.0"
description = ""
authors = ["Leonardo Gregianin <leogregianin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
google-api-python-client = "^2.49.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.1"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
